Move parts of the session code to sansio This will allow it to be used in Quart, thereby reducing duplication, and ensuring the APIs match.
